Rename Path Generator to Resource Path Naming Strategy
