+++ src/Naming/ResourcePathNamingStrategyInterface.php
@@ -9,12 +9,21 @@
  * file that was distributed with this source code.
  */
 
-namespace ApiPlatform\Core\Routing;
+namespace ApiPlatform\Core\Naming;
 
 /**
+ * Generates a path from a resource name.
+ *
  * @author Paul Le Corre <paul@lecorre.me>
  */
-interface ResourcePathGeneratorInterface
+interface ResourcePathNamingStrategyInterface
 {
+    /**
+     * Generates the base path.
+     *
+     * @param string $resourceShortName
+     *
+     * @return string
+     */
     public function generateResourceBasePath(string $resourceShortName) : string;
 }
